requirements.txt
streamlit>=1.31.1
streamlit-aggrid>=0.3.4
pandas>=2.1.4
psycopg2-binary>=2.9.9
requests>=2.31.0
tqdm>=4.66.2
plotly>=5.0.0
python-dotenv>=1.0.1

